api/requirements.txt
flask==1.1.2
flask-cors==3.0.10
flask-bcrypt==0.7.1
flask-jwt-extended==4.3.1
requests==2.27.1
peewee==3.14.8
python-dotenv==0.19.2
psycopg2-binary==2.9.2
packaging==20.9
pkce==1.0.3
urllib3==1.26.8
discord==1.7.3
google-api-python-client==2.37.0
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0
google-auth==1.29.0
googleapis-common-protos==1.54.0
google-pasta==0.2.0
itsdangerous==1.1.0
spotipy==2.19.0
cryptography==35.0.0